pyproject.toml
[tool.poetry]
name = "bp-training"
version = "0.1.0"
description = ""
authors = ["ben-pollard <benpollard0@googlemail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
mlflow = "^2.17.2"
transformers = "^4.46.2"
datasets = "^3.1.0"
torch = {version = "^2.5.0+cpu", source = "pytorch_cpu"}
evaluate = "^0.4.3"
pytorch-lightning = "^2.4.0"
seqeval = "^1.2.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
dagster = "^1.9.2"
dagster-webserver = "^1.9.2"
dagster-mlflow = "^0.25.3"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.5"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.dagster]
module_name = "workflow.definitions"
code_location_name = "workflow"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
